requirements.txt
# Generic
librosa==0.9.2
matplotlib
more-itertools
numpy==1.22.*
packaging
pandas==1.3.*
pathos==0.2.*
pyparsing==2.4.7
rich
scikit-learn
scipy==1.8.*
seaborn
soundfile
typer==0.4.*
wandb==0.12.*

# Code quality
black
black[jupyter]
flake8
flake8-annotations-complexity
flake8-cognitive-complexity
flake8-docstrings
flake8-simplify
isort[colors]
mypy
pydocstyle
pylint

# PyTorch & Hydra
--find-links https://download.pytorch.org/whl/cu113/torch_stable.html
torch==1.11.*
torchaudio==0.11.*
torchmetrics==0.9.3
torchvision==0.12.*
hydra-core==1.2.*
git+https://github.com/pytorch/hydra-torch/#subdirectory=hydra-configs-torch
git+https://github.com/pytorch/hydra-torch/#subdirectory=hydra-configs-torchvision
pytorch-lightning==1.5.*
lightning-bolts==0.4.*
pytorch-yard==2022.9.1
torch_tb_profiler==0.3.*
torchinfo==1.5.*
hypnettorch==0.0.4
auraloss==0.2.*
cdpam==0.0.6
pystoi==0.3.3
pesq==0.0.4
random-fourier-features-pytorch==1.0.1


# Encoders
transformers==4.41.0
pyworld==0.3.4
